requirements.txt
PyMySQL==1.0.3
python-dotenv==1.0.0
types-PyMySQL==1.0.19.7